Refactor participant table
